Fixed - do health check before device restoration
